✨ (App.jsx, MicrosoftPortalsShadcn.jsx): refactor Microsoft Portals tool by renaming component, removing deprecated implementation, and enhancing functionality with new modular components for improved user experience.
